BUG: fix paths for Mac build running helper apps
On mac, the '.app' approach typically does not require special
path settings for libraries, but for the use case of these CLI
modules where they expose custom interfaces to built in slicer
CLIs, they need to be able to reliably find the appropriate executable
corresponding to the slicer instance that launched them.

For this, we rely on the SLICER_HOME env variable, which is set
for all platforms and uniquely defines where to look for the
helper CLIs.