ENH: Remove hardcoded relative path to executables in Slicer
Paths to executables that need to be found in Slicer were hardcoded (eg: lib/Slicer-4.4/cli-modules) which was creating errors when the version of Slicer was modified. The developer had to think about changing that path every time this was happening. I replaced this hardcoded path with the constant "Slicer_CLIMODULES_BIN_DIR" that is defined in "vtkSlicerConfigure.h" which is configured in the project of Slicer (this file is at the root of the inner-build directory of Slicer). The inner-build directory of Slicer is defined as "${Slicer_HOME}" in SlicerConfig.cmake .